+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # `versions` Changelog
 
+## Unreleased
+
+#### Changed
+
+- **Breaking:** `Mess::chunk` renamed to `Mess::chunks` to match `Version`.
+- **Breaking:** `Mess` now stores smarter `Vec<MChunk>` instead of `String`s.
+
 ## 1.0.1 (2020-06-15)
 
 #### Changed